Update JavaDoc documentation
This primarily updates the JavaDoc documenation within the compiler.
This process is not yet complete, though coverage is pretty good at this
time.

In addition, this moves the package wyc.type into wyil.type.  This helps
to reduce the amount of stuff in the wyc package, and increace the
amount in the wyil package.  Furthermore, this is somehow a "low-level"
package which fits nicely in wyil.