Add new opcodes and type new expressions properly
This adds new opcode (staticnew and staticreference) for representing
new expressions and reference types without lifetimes.

In addition, update the new operator to properly generate a type based
on its lifetime.  Previously, it was just returning a static lifetime
for everything.

Some lifetime tests are now passing, though not all.  This is because
only the static lifetime and this lifetime are currently considered.