Partial implementation of method selection #764
This provides a partial implementation of the algorithm for method
selection in the presence of lifetime parameters.  This is a non-trivial
algorithm because of the non-trivial number of lifetime bindings which
can be generated.

At this stage, the algorithm has some bugs and is not correctly
generating the "most precise" type.  In particular, it is preferring
smaller lifetimes to larger ones.