Support lifetime args; Fix selectCallableCandidate
This adds support for providing explicit lifetime arguments at the
source level.  For example, writing an invocation like "m<this>(x)",
etc.

This also fixes some problems with the way selectCallableCandidate was
working in the presence of "similar" types.  These are types which are
raw subtypes of each other, but actually differ in some way (e.g. from
different declarations, or have type invariants, etc).  The selection
procedure can now continue in the presence of such types and still find
an outright winner if one exists.