Fix for #817
This implements a straightforward fix for #817.  The essential problem
was that it was mixing the generation of WyAL expressions and the
assignments for a multiple assignment statement.  Instead, it must be
done in a strict order.  All the rhs expressions must be generated
before any of the assignments are applied.