Remove Type.Intersection and Type.Difference
This partially implements a significant change to the way that type
checking occurs.  Specifically, Type.Intersection and Type.Difference
are removed and, hence, no longer consider part of a syntactic or
source-level type.  Instead, they are replaced with a separate concept
called the SemanticType which effectively maintains a combination of
types in DNF.  Thus, we can perform subtyping checks using these
SemanticTypes, but keeping all recorded types as just Type.