SemanticType & FowardPropagation
This means that SemanticType is now properly in the WhileyFile
hierarchy, along with the other types and expressions.  This is helpful
because it allows e.g. visitors to be used with them seemlessly.  It
also means they could potentially be stored into a WyilFile if needed
(e.g. to support incremental computation).

The FlowTypeChecker now employs a (slightly broken) form of type
propagation.  This propagates type information from the destination of
a given expression down the AST making up that expression.  This allows
one to determine the appropriate representation at each point.  In
addition, it means we can compute the definitive (concrete) type for
each expression.  Hopefully this solves a set of long standing issues
related to the translation of Whiley for low-level backends.  Fingers
crossed!!

This also makes a "best effort" update for test cases broken by the removal
of effective types.  In fact, there were surprisingly few test cases
broken.  Of these, all but four were easily resolved.  The remaining
four indicate problems in the current mechanism for extracting types.

As an aside, the paucity of broken tests probably indicates this feature
was relatively poorly tested.

This adds support for type checking a forward's direction with multiple
possible type cases.  This allows resolving invocations by passing
forward type cases, rather than individual expected types.  Type cases
are then narrowed down based on their return type.

Have reorganised FlowTypeChecker a little and this is looking better.
Split out FlowTypeUtils.

This changes the checking for runtime type tests to use the strict
subtype operator, rather than the relaxed operator.  Something tells me
there are further things to work out here, since relaxed probably
shouldn't be used in some cases where it currently is.