Design for ReadWriteTypeExtractor
The current design doesn't appear to work, and seems to be related to
the lack of conversion to DNF.  Potentially, taking combinator style and
applying it to AbstractTypeExtractor ... and that might do it.

This reverts the calculation of array and record lvals to use
checkExpression.  This is now working as expected, though I'm sure there
are edge cases to figure out.

Remove Type.asCallable() and related methods (e.g. asArray())