Update test cases #854
With a little bit of help from sed, I have now updared the test cases to
the new syntax.  Have also include "_" following Java as I quite like
this sometimes.  That means, for example, that 0b0101_1110 is a valid
binary literal.  I have also included "_" for integer literals as well.
So, that means "1_000" is a valid integer literal.