Add WyllFile and accompanying compile task
This now restores the WyllFile and begins the process of fixing up the
old compile task.  The essential idea is that a wyll file provides a
C-like imperative language which provides a low-level target for
compiling Whiley to.  This can then be used as a bouncing board to other
low-level targets, such as C/Java/JavaScript/LLVM/EVM, etc.

The command "llcompile" will be used to compile a whiley file to wyll
file; a similar command will provided for decompiling wyll files.