Add  Wyll Interpreter
The interpreter executes WyLL bytecodes.  The idea is simply to test the
translation from WhileyFile to WyllFil.  It is unclear to me at this
time whehther or not this interpreter will be required in the long run.
For now, it's providing a useful way of testing the LowLevelCompileTask.
In the future, I may replace this interpreter with a "Universal Machine"
which provides a simple hardware abstraction.  That would be useful as a
launching pad to machine code.

This also makes various improvements to the LowLevelCompileTask and a
good number of tests are now passing with the interpreter.  There's a
lot of work still to do, however, on runtime type tests, coercions and
other bits of the jigsaw.