Updated to WyCC feature/7-build
This updates the code to compile against the feature/7-build
branch of WyCC which is now in develop.  In particular, this means
implementing the Build.Platform which is now controlled within WyCC.

This also updates of the test harness after removal of the various
Compile / Run commands.  I wonder though if we're going to go through
with all of this!!