Working on NameResolution
The NameResolution phase is starting to work.  However, need now to
actually take the resolved names and apply them to e.g. Type.Nominal,
etc.  This means we need to turn the resolved name into the
corresponding Declaration.  For local declarations, this should be easy
enough.  However, for non-local declarations we will need to find a way
to import them from the respective modules.