Bug for interpreter invocation
There was a bug in the interpreter related to executing callables
declared in dependencies.  The problem was that I was looking up based
on the Type.Callable signature, but this doesn't actually work anymore.
Specifically, because two Type.Callable's will not compare equal unless
they refer to the *same* Decl.Named instance ... and if one is a stub
declared in the local module, and the other is declared elsewhere ...
problem.