QuickCheck: is generating primitive values
At this stage, the quick check algorithm is now generating some input
values.

This adds support for array and records, and also for nominal types.

This adds a framework for running the quick check command on the invalid
test cases.  At this stage, not much happens...

The quick check tool is starting to work on the invalid test suite.
However, I need to update the interpreter to ensure syntactic exceptions
are thrown.  Furthermore, I might need to update the intepreter to
ensure sufficient runtime checks are being performed.

This updates the interpreter to catch some more runtime errors.
However, there are still quite a few which are unaccounted for.

This adds support for checking type invariants to the interpreter.
Unfortunately, this has a identified a whole truck load of bugs in the
existing system which are not being caught by verification (or perhaps
are ignored).

This puts in place some improvements to the testing framework.

This adds support for generating byte types.
This adds support for union types.  Also tidies up the DomainWrapper.
Adds support for limiting recursive types and paves the way for
configurable bounds, etc.