Various fixes to QuickCheck & Interpreter
This implements various fixes to the tool itself, as well as to the interpreter.
Unfortunately, the biggest issue remaining is performance which is relatively
hard to fix.  This also adds JModelGen to the pom.xml.

This adds support to QuickCheck for handling reference types.  Furthermore, this
introduces the concept of "aliasing width" to handle references to different
objects with the same value.

This adds support for synthesising lambdas based on.  In principle, we
could also look through existing functions for matching lambdas.