Allow skipping methods in QuickCheck
QuickCheck now, by default, skips methods.  This can be overridden if it is the
desired behaviour.  For example, in the standard library, it probably is the
desired behaviour.  However, in the case of the WyBench suite it isn't because
methods provide the outer loops which tend to be quite expensive.