Working on extracting stack frame information
Extracting stack frame information would be very useful for error reporting.
This encodes the frame information into the WyilFile as contextual information
on the SyntacticMarker.  However, there is currently no way to actually this
information as it doesn't get printed by the default Build command.  It's not
clear to me how best to handle this right now.