Stripped out SemanticType #929
This strips out SemanticType altogether and updates various algorithms
involved.  In particular, have deleted a bunch of algorithms, such as
ConcreteTypeExtractor.  Have made other simplifications.  At this point,
it does not compile, though is close.

This incorporates StrictIncompleteSubtypeOperator from WyJS and simplifies the
code base significantly.  At this stage, everything compiles but a larger number
of tests are failing.  At least one significant issue is immediately apparent:

> nat i = 0
> i = i + 1

This fails to type check because int <: nat no longer holds.  It's not clear to
me at this time what the right option is.  We could make arithmetic operators
type polymorphic or something weird like that.

This puts through a lot of improvements to the subtyping implementation.  In
particular, only 22 test cases remain.  There are still some issues though.