Fix #968
This puts through a simple fix which dramatically improves the performance of
compilation.  Specifically, by removing the RecursiveTypeAnalysis phase.  This
may seem a bit radical but, basically, it wasn't being used.  It was possible to
fix it using a variation of Tarjan's algorithm for finding strongly connected
components, but what's the point if its not being used?