Add WyilFile.Type.Selector
This adds the concept of a "type selector" which produces a refinement
of a given type based on which tag values are known to be possible.  For
example, given a type int|null the selector (*|_) returns the type int.
Type selectors can also be used for printing more concise information
about a given type.  For example, printing "{..., int f}" rather than
"{int a, int b, int f}" if that is relevant.

At this stage, type selectors are complete upto recursive types.
However, obviously, support is required for recursive types now.