Implement Type.Selector generator
This implements a very simplistic generator for type selectors.  It's simplistic
because it essentially just applies the strict subtype operator, and does not
attempt to generate a true morphism, etc.  However, it does work for now and
paves the introduction of Type.Selectors in production.  Furthermore, it opens
up the very real possibility of forming type refinements using selectors.