Remove Multiple Expressions #975
This removes the aged and awkward concept of "multiple expressions".
That is, special cases for handling multiple return values from
invocations.  The key to this is the reintroduction of tuples and a new
concept of "shape".  The intuition behind type shae is that it
identifies the number of "dimensions" of a type.  For example, an atom
has one dimenions whilst a tuple has (strictly) two or more.  Likewise,
void has zero dimensions.

The language syntax has changed and now follows Rust in requiring clear
indicators for tuple lvals.  Specifically, you must enclose the assigned
variables in braces to indicate a tuple lval.  This has simplified
things a lot.  For example, every expression now has only one return
type.  Likewise, a function type takes a given type and maps it to
another type (i.e. rather than a tuple of types to a tuple of types).

In addition, we now have a tuple initialiser expression which is
again indicated using braces.  It remains to reintroduce tuple
initialisers and, likewise, the tuple access operator.

At this stage, verification is working again but it is really in
trouble.  In general, I doubt it will be able to handle the full
reintroduction of tuples.  However, it's the best I can do given the
circumstances.