Support multiple variable declarations #979
This adds support for multiple variable declarations of the form:

> int x, int y = swap(a,b)

An interesting alternative syntax would be:

> (int x, int y) = swap(a,b)

However, parsing this syntax is definitely more involved due to the open
brace potentially conflicting with the started of a bracketed type (e.g.
"(&int)[]").  Therefore, have stuck with the simplest option at this
stage!

This additionally adds some new test cases for tuples and tuple
initialisers.