Initial Design for New FlowTypeChecker
This adds a preliminary implementation of the propsed new flow type checking
algorithm.  The ultimately aim is to allow forward propagation so that, for
example, values are aligned with the declared type of the target as much as
possible.  For example, consider the expression:

js::core::string s = "Hello World"

Currently, the type of the string literal is simply int[].  If we want
js::core::string to be a native JavaScript string this means we construct a
valid int[] and then later coerce it into a native string.  With forward
propagation enabled the type of the string literal would be js::core::string and
we could generate a native string directly.

In addition, forward propagation enables disambiguation based on the return type
of a method or function invocation.  This is particularly useful in the presence
of templates, such as when a templated function accepts no arguments or no
arguments of the template type.

The design as it is in my head, at this stage, is that we first perform a
backwards propagation which may result in one or more valid typings based on the
"natural" type of expressions.  During this stage, most type errors would be
caught.  Then, we perform a forward propagation to push assigned types as far
down the tree as possible, noting that we cannot always go all the way down.
Observe, that I believe it's possible to do the whole thing in one pass, but
actually this going to very hard to implement coherently.  Hence, splitting out
seems like a reasonable compromise.

This puts through support for working with multiple typings simultaneously.  The
essential idea is that we progressively rule out typings until either "the one"
typing remains, or no typings remain (indicating a subtyping error), or multiple
typings remain (indicating some ambiguity).

At this stage, the code handles multiple typings through the concept of a
Typing.Vector which provides the type for a given variable (i.e. expression) in
each active dimension.  We can then project over this vector to, for example,
extract all the array element types.