Update for Reference Semantics #985
This puts through a significant update to the implementation of
references in Whiley which offers backends something which is closer to
the machine.  This allows, for example, native references in JavaScript
to be represented directly in Whiley.  There are still some rough edges
around this, however.  But, mostly it works.