Backwards type inference works upto templates!
The improved algorithm for backwards type inference is now working well, and
enabling new code to compile which was previously impossible.  However, it
remains to handle templates and some aspects of lifetime parameters.

This redesigns the API again to introduce the concept of a Typing.Environment
along with a set of Typing.Constraints.  This is going to help with the binding
process very shortly now ...

There are a lot of problems at this stage, though things are slowly
starting to work.  Perhaps the main issue is that the code is *serious*
ugly at this point.  There are quite a few ways it could be improved,
and clearly I'm going to have to work on that.

At this stage, binding is starting to work though it's pretty clear
we're going to need lower and upper bounds.  Also, void arrays are
currently a major problem for some reason.