isSubtype now returns Constraints
This updates isSubtype to return a set of constraints, rather than a boolean
yes/no.  Thus satisfiability is subject to the constraints being satisfiable.
This allows the bind() algorithm to be merged with isSubtype() to avoid
repeating the same thing.

This also reintroduces the "Any" type.  This allows variable boudns to be
represented initially using Type.Any :> Type.Void which avoids, for example,
using null to signal no solution is found.