Updated Type Inference Algorithm
This puts through a useful update to the core type inference algorithm
which now attempts to keep "clear winners".  That is, typing solutions
which subsume other typing solutions.

In addition, the algorithm now solves for the "most precise" typing
solution.  That is, the least solution.  This required some simple fixes
to make it work, compared with before.