Beginning Strict Subtyping #998
This introduces the use of existentials in specific places, such as for
constants, array initialisers and record initialiser.

Update test cases; fix parsing of Expr.New

Previously parsing of Expr.New was broken because it didn't account for
lifetimes correctly.  This is now fixed.

Add concept of "ancestors" for types

This introduces the concept of an ancestor relationship between type
declarations."