Add leastUpperBound/greatestLowerBound #998
Have put through much of the infrastructure for these methods, but need to
actually implement them.

It's starting to feel like it's working, and it makes sense.

The concept of an "alias" is also emerging as something which would be handy.

This puts through various bug fixes.  There are some interesting outstanding
problems with references.

At this stage, there remain a number of unresolved questions around constraint
solving.