Support bounded existentials
This puts through better support for representing initialisers.  For
example, consider "nat x = 1".  Instead of typing this as just "int :>
?0; nat :> ?0" we type as "?0 :> ?1; nat :> ?1; nat :> ?0".  In this
case it doesn't make much difference, but in other cases it most
definitely does.  The key problem is that the constraint solving
algorithm is not powerful enough, even when the constraints are
solvable.

Working on improved constraint solving

Am working towards enabling constraint solving to employ backtracking.
It's not completely clear to me whether or not this is the absolute best
solution, but I don't see any other way.

At this stage, have tidied up the code base.  It's not completely
working yet because greatestLowerBound(Type.Record,Type.Record) is
broken for open records.  Otherwise, it's looking better.

The key is that, as before, the constraint set carries around an
candidate solution to the constraints.  However, if this becomes
unsatisfiable, it does not immediately mean the whole thing is.
Instead, we need to recompute from scratch using backtracking to be
sure.

The algorithm for this thing is really getting quite complex.  I need to think
very carefully about whether or not I need it to be *this* complex.

This puts through an attempt to implement forward propagation.  There are a
number of unknowns at this stage, and it's going to require some big changes.