Working on forward type propagation
This forward type propagation has some potential advantages related to concrete
information.  At this stage, it is starting to work.  However, there are some
strange errors cropping up.

The concept of array projection is rather useful in the overal type inference
algorithm.  At this stage, it seems to work.  However, I suspect it will need
further refinements, etc, in the future.

Next on the list is to fix reference and record inference.  I haven't figured
out how best to handle equality at this stage, but anyhow.

Notes:

* Invoke must be next.

* Need to figure out how to manage the inverse projections.  Two options:
  introduce special types (e.g. to access base type, or to apply projection).
  Second, create different subtyping constraints.

This removes the counter for the number of existentials allocated from
the Constraint.Set and places it into the Typing.  This actually offers
several performance improvements and reduces the API for Constaint.Set.