Add Subtype Constraint Forms
Processing of delayed constraints is pretty tricksy.  At this stage, things are
starting to work upto function overloading.

Currently, main issues:

(1) cannot propagate in both directions across a projection constraint.  This is
actually severely limiting.

(2) working on overloading.  Key issue is associating meta-data with expressions
so we can actually type them in the end.

Currently there is a problem with the creation of type frames and
equality.  The issue arises because equality creates a type variable
which has no solution until after the pushTypeFrame has completed.  But,
for invocations, pushTypeFrame runs a projection which, in turn, flushes
the stack.  But, what we need is to flush the stack whilst performing
the projection (or something).

Basically, separating out frame allocation from variable allocation
provides greater flexibility.

Have added Subtyping.ProjectionConstraint and used this for array
initialisers.  Presumably, it will need to be used for record
initialisers, tuple initialisers, etc.

A key outstanding question is the issue of delayed constraints and
solution satisfiability.  Do we really need every variable to have a
solution to be satsified?  Do we really need an exponential search or
can we just cap solutions and proceed?  These are the big questions.

Currently, arrays of arrays are failing for reasons unknown.

More stuff is starting to work after several big refactorings:

(1) Folding the whole system into using Subtyping.Constraints rather
than having two separate systems.

(2) Moveing towards a forwards propagation model.

However, the big outstanding problem is that of deciding when a
contraint set is satisfiable or not.  Currently, the reliance on every
varaible having an exact solution is really limiting.  All I really need
at any given moment is to know whether or not there is actually a
*contradiction*.