Change of tack
Have pulled out the complex isSatisfaction() algorithm in favour of
something much simple.  Using this, am attempting to get all constraints
to pass at least.  Then I'll improve the algorithm.

Currently, struggling to understand problem with the projection
constraint.  It's not working for constrained records ...

One of the key challenges here is that I need to decide very carefully how this
scheme would work in the context of type variables arising from templates.  In
particular, we can have type variables flowing up and flowing down.  Type
variables need to be narrowed, whilst other variables don't.  How do we
reconcile this?

the new approach will work better in all cases where there are no type
parameters being inferred.  Then, that will still leave a bunch of problematic
issues there, but at least it's more localised.

Things are starting to work, though array initialiser remain a problem (no
surprises there).  The API probably might be improved if we moved from
Type.Variable instances to int + Typing instances.

Most tests are now failing because of templates, so this has to be the
priority.  In general, the code base needs a fair bit of tidying up.
This will probably help to improve quality as well.

Currently, there is a problem in how we deal with expressions which don't need
to have a return type.  Previously, I used the "hasReturn" boolean but this
doesn't work for lambda's where we don't know whether or not it has a return.

At this stage, have removed hasReturn but this means some tests now failing
which weren't previously.  One example is Array_Valid_9 which has a standalone
invocation.

Am attempting to replace pullArrayInitialiser with filter().  However,
it's not exactly working.  In order to support type refinement at
assignments, we need to pull down a refinement.  The problem is figuring
out what the right refinement is!!

Lots of work on errors

There are still a number of outstanding issues:

(1) Projecting forward doesn't consider choices.  E.g. (&int)|(&bool)
x = new 1 fails

(2) Current approach to filtering fails for lambdas.  Greatest Lower
Bound may help?

(3) Setting non-bool type on a condition doesn't make sense

(4) type variable projections (e.g. for arrays)

(5) Typing (int[])|(bool[]) x = [1] gives that same type to the array
initialiser.  This doesn't make sense, surely?

(6) Typing indirect invocations.  This fails because cannot easily
determine supertype of all callable types.  Would be handy to do this!
inference.  Then, perhaps the stricter approach to subtyping.

Have added a method "disjoint()" which determines whether two types are entirely
disjoint or not.

The remaining outstanding problems relate exclusively to the fundamentally
problematic ways in which the pull and pushing is working.  This is especially
true for initialisers, etc.