Redesigned Push / Pull Mechanism
The new strategy is based around having separate methods for push and pull.
That is, instead of a single method (e.g. inferArrayInitialiser), we now have
two methods (pushArrayInitialiser, pullArrayInitialiser).  In some sense, this
results in more code as we end up with some duplication between the two methods.
However, instead of one larger and less coherent method, we utlimately end up
with two smaller and far more coherent methods.  In fact, this has helped a lot
to apply key optimisations and understand exactly where we want to push versus
pull.

In the end, the strategy appears to have paid of handsomely.  In particular, it
has clarified that we don't in fact need any more specific constraint forms (as
I was originally assuming).  For example a constraint to indicate, something
should have an underlying type.  This is two reasons.  Firstly, using those
constraint forms in a pull method is unnecessary as we can always use the
underlying type in the subtype position (i.e. because there is no constrained
type in the supertype position).  Secondly, when pushing a template parameter it
turns out (rather surprisingly) that we're in the same position.  This is
because we push through return values meaning that templates are either
established from returns, or are essentially "open" (i.e. using an underlying
type in the subtype position is OK).  There I think a few cases where this does
fail, but they are few and far between.