Detect Non-Export/Public Tests
This adds static detection for test cases which are not exported or public.
Such test cases will pass the WyC tests here, but will break for other backends.
Hence, detecting them here makes life easier.

As a result of this detection, several test cases have been changed to confirm
properly with the requirement that test methods are public and exported.