Fix #1040
This puts through a relatively simple fix for type inference of
templates at function invocation sites.  That this was missing before
somehow went unnoticed.  The fix was a little hacky, but basically
amounted to adding an extra type variable to be inferred for callable
frames.  Some tricky edge cases, but we got there in the end.