Fix #1045
This puts in place a rather hacky fix for this issue.  The problem is
that I have no way to write a reference into the syntactic heap as an
instance of Value.  This is necessary as the stack frames are stored
accordingly.  Therefore, I have simply written the result as a string!