Towards Fixing QuickCheck
The QuickCheck command is now operational, though its not working very well.
The problem is that the command cannot determine the source artifacts for the
WyilFile stored on disk.  This is because the command is executed after the
build, and this causes the WyilFile to be loaded from disk.  Since it doesn't
store the list of source artifacts on disk, this list is empty when it is
loaded.