Updates for #1120
This puts through various updates for checking static variable accesses in pure
contexts.  In particular, should not be able to access a non-final static
variable anywhere except an impure context.  Have updated FunctionalCheck and
also a number of test cases.

There is one remaining issue, in that there's no reason why we cannot access a
static variable within an assert or assume statement (itself within a method).
At the moment, this is ruled out.