Implement TestFile
This puts through an initial implementation of the file format for tests, along
with a parser.  Furthermore tests are run and the generated errors compared
against what was expected.  This required extracting error codes from
SyntaxError instances.  Have added a few flags, such as the following:

* build.whiley.ignore.  This can be used to ignore a test.
* WhileyCompiler.issue.  This can be used to register a github issue against
  this test.

And there are a few others as well.  Overall, its working nicely.