Compiles and runs tests
At this stage, it now compiles and runs tests.  However, it is failing
automatically if the test fails.  Its not checking the error codes.

Furthermore, the method being executed is always in a wyil file named after the
test file, which is not ideal.  Better to have it be a default of "main" or
something.