Working on error markers
Need to parse the markers coming out of the Whiley compiler.  To be honest,
ideally, it would be nice not to have to parse anything.  Anyways, for now, I
guess that's the easiest option.

Once I've parsed them I should be able to sort them and then just do a linear
comparison.