Convert valid and invalid tests over
This adds all the original tests in the valid suite as tests in the new
testfile format.  In each case, we simply pipe the entire test file into
a single file and expect no errors.  Meta data is included for tests
which were previously marked ignored.

For the invalid tests, things were a bit more complicated.  Many of the tests
were actually broken, and needed to be fixed.  Others were missing test methods
meaning they could not be executed, etc.  After a fair amount of manual work,
they are all convered over.