Implement #1139
This adds support for detecting tests that are marked as ignored, but actually
now pass.  The solution is not full proof though.  In particular, the test must
be correctly specified for it to work.  For example, if the test is missing
error messages which are expected then the tool won't notice its actually
passing (how could it?).